requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
tensorflow>=2.8.0
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.3
streamlit>=1.0.0
mlflow>=1.20.0
boto3>=1.26.0
botocore>=1.29.0
matplotlib>=3.4.3
seaborn>=0.11.2
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
black>=22.0.0
flake8>=6.0.0
python-dotenv>=1.0.0
mkdocs>=1.4.0
mkdocs-material>=9.0.0
terraform>=1.0.0
infracost>=0.10.0
tfsec>=1.0.0
pyyaml>=6.0.0
jinja2>=3.1.0
requests>=2.28.0
mypy>=1.0.0
types-requests>=2.28.0
types-boto3>=1.26.0
isort>=5.10.0
pylint>=2.15.0 